dailybuild:
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7549 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/HumanInterface/HILeap.cpp
@@ -40,9 +40,9 @@ void HILeap::Update(float dt) {
 		LeapFrame frame = GetLeapFrame(leap, 0);
 
 		// Skeletonの不足分を用意
-		PrepareSkeleton(frame.leapHands.size());
+		PrepareSkeleton((int)frame.leapHands.size());
 
-		for (int h=0; h<frame.leapHands.size(); ++h) {
+		for (int h=0; h<(int)frame.leapHands.size(); ++h) {
 			LeapHand hand = frame.leapHands[h];
 
 			HISkeleton* skel = skeletons[h]->Cast();
